requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.5
django-ckeditor==6.7.2
django-import-export==4.3.4
django-js-asset==3.0.1
djangorestframework==3.15.2
filelock==3.16.1
pillow==11.1.0
platformdirs==4.3.6
pynput==1.7.7
six==1.16.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
virtualenv==20.27.0
